+++ feature/insertRoutine/src/androidTest/java/routiner/feature/insertRoutine/ExampleInstrumentedTest.kt
@@ -1,12 +1,10 @@
-package com.hegunhee.routiner.insertRoutine
+package routiner.feature.insertRoutine
 
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
@@ -19,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.hegunhee.routiner.insertRoutine.test", appContext.packageName)
+        Assert.assertEquals("routiner.feature.insertRoutine.test", appContext.packageName)
     }
 }
