+++ feature/insertRoutine/src/main/java/routiner/feature/insertRoutine/common/InsertCommonScreen.kt
@@ -1,4 +1,4 @@
-package com.hegunhee.routiner.insertRoutine.common
+package routiner.feature.insertRoutine.common
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
@@ -18,8 +18,8 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
-import com.hegunhee.routiner.insertRoutine.R
-import com.hegunhee.routiner.insertRoutine.daily.NormalTextField
+import routiner.feature.insertRoutine.R
+import routiner.feature.insertRoutine.daily.NormalTextField
 
 @Composable
 internal fun ColumnScope.RoutineTextEnterScreen(
