+++ feature/insertRoutine/src/main/java/routiner/feature/insertRoutine/daily/InsertRoutineRootScreen.kt
@@ -1,4 +1,4 @@
-package com.hegunhee.routiner.insertRoutine.daily
+package routiner.feature.insertRoutine.daily
 
 import android.content.Context
 import android.widget.Toast
@@ -36,12 +36,12 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.repeatOnLifecycle
-import com.hegunhee.routiner.insertRoutine.R
-import com.hegunhee.routiner.insertRoutine.common.CategoryDescriptionScreen
-import com.hegunhee.routiner.insertRoutine.common.CategoryLazyList
-import com.hegunhee.routiner.insertRoutine.common.CategoryTextEnterScreen
-import com.hegunhee.routiner.insertRoutine.common.RoutineTextEnterScreen
+import routiner.feature.insertRoutine.common.CategoryDescriptionScreen
+import routiner.feature.insertRoutine.common.CategoryLazyList
+import routiner.feature.insertRoutine.common.CategoryTextEnterScreen
+import routiner.feature.insertRoutine.common.RoutineTextEnterScreen
 import kotlinx.coroutines.flow.SharedFlow
+import routiner.feature.insertRoutine.R
 
 @Composable
 fun InsertRoutineRootScreen(
