+++ feature/insertRoutine/src/main/java/routiner/feature/insertRoutine/daily/InsertRoutineUiEvent.kt
@@ -1,4 +1,4 @@
-package com.hegunhee.routiner.insertRoutine.daily
+package routiner.feature.insertRoutine.daily
 
 sealed interface InsertRoutineUiEvent {
