+++ feature/insertRoutine/src/main/java/routiner/feature/insertRoutine/daily/InsertRoutineViewModel.kt
@@ -1,67 +1,56 @@
-package com.hegunhee.routiner.insertRoutine.repeat
+package routiner.feature.insertRoutine.daily
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import routiner.core.domain.usecase.category.DeleteCategoryUseCase
 import routiner.core.domain.usecase.category.GetCategoriesFlowUseCase
 import routiner.core.domain.usecase.category.InsertCategoryUseCase
-import routiner.core.domain.usecase.date.GetSortedDayOfWeekListUseCase
-import routiner.core.domain.usecase.repeatRoutine.InsertRepeatRoutineUseCase
+import routiner.core.domain.usecase.routine.InsertRoutineUseCase
+import routiner.core.util.getTodayDate
 import dagger.hilt.android.lifecycle.HiltViewModel
 import routiner.core.model.Category
-import routiner.core.model.DayOfWeek
-import routiner.core.model.RepeatRoutine
+import routiner.core.model.Routine
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asSharedFlow
-import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class InsertRepeatRoutineViewModel @Inject constructor(
+class InsertRoutineViewModel @Inject constructor(
     private val getCategoriesFlowUseCase: GetCategoriesFlowUseCase,
     private val deleteCategoryUseCase: DeleteCategoryUseCase,
     private val insertCategoryUseCase: InsertCategoryUseCase,
-    private val getSortedDayOfWeekListUseCase: GetSortedDayOfWeekListUseCase,
-    private val insertRepeatRoutineUseCase: InsertRepeatRoutineUseCase,
+    private val insertRoutineUseCase: InsertRoutineUseCase,
 ) : ViewModel() {
 
     private val categories = getCategoriesFlowUseCase()
 
     private val selectedCategory: MutableStateFlow<String?> = MutableStateFlow(null)
 
-    private val _dayOfWeek: MutableStateFlow<List<DayOfWeek>> =
-        MutableStateFlow(getSortedDayOfWeekListUseCase().map {
-            DayOfWeek(date = it, isSelected = false)
-        })
-    val dayOfWeek: StateFlow<List<DayOfWeek>> = _dayOfWeek.asStateFlow()
-
-    val uiState: StateFlow<InsertRepeatRoutineUiState> =
+    val uiState: StateFlow<InsertRoutineUiState> =
         categories.combine(selectedCategory) { categories, selectedCategory ->
             if (selectedCategory == null) {
-                InsertRepeatRoutineUiState.Categories(categories, listOf())
+                InsertRoutineUiState.Categories(categories)
             } else {
-                InsertRepeatRoutineUiState.Categories(categories.map { category ->
+                InsertRoutineUiState.Categories(categories.map { category ->
                     if (category.name == selectedCategory) category.copy(isSelected = true)
                     else category
-                }, listOf())
+                })
             }
-        }.combine(dayOfWeek) { categories, dayOfWeek ->
-            InsertRepeatRoutineUiState.Categories(categories.categories, dayOfWeek)
         }.stateIn(
             scope = viewModelScope,
-            initialValue = InsertRepeatRoutineUiState.Init,
+            initialValue = InsertRoutineUiState.Init,
             started = SharingStarted.WhileSubscribed(5_000)
         )
 
-    private val _uiEvent : MutableSharedFlow<InsertRepeatRoutineUiEvent> = MutableSharedFlow()
-    val uiEvent : SharedFlow<InsertRepeatRoutineUiEvent> = _uiEvent.asSharedFlow()
+    private val _uiEvent: MutableSharedFlow<InsertRoutineUiEvent> = MutableSharedFlow()
+    val uiEvent: SharedFlow<InsertRoutineUiEvent> = _uiEvent.asSharedFlow()
 
     fun onClickCategory(categoryName: String) {
         if (categoryName == selectedCategory.value) {
@@ -83,38 +72,28 @@ class InsertRepeatRoutineViewModel @Inject constructor(
     fun onClickCategoryInsert(insertCategoryText : String) {
         viewModelScope.launch {
             if(insertCategoryText.isBlank()) {
-                _uiEvent.emit(InsertRepeatRoutineUiEvent.InsertCategoryNameEmpty)
+                _uiEvent.emit(InsertRoutineUiEvent.InsertCategoryNameEmpty)
                 return@launch
             }
             insertCategoryUseCase(Category(insertCategoryText))
         }
+
     }
 
-    fun onClickDayOfWeek(dayOfWeek : String) {
-        _dayOfWeek.value = _dayOfWeek.value.map {
-            if(it.date == dayOfWeek) it.copy(isSelected =  !it.isSelected)
-            else it
-        }
-    }
-
-    fun onClickRoutineInsert(routineName: String,selectedDayOfWeek : List<String>) {
+    fun onClickRoutineInsert(routineName: String) {
         viewModelScope.launch {
             if (routineName.isBlank()) {
-                _uiEvent.emit(InsertRepeatRoutineUiEvent.RoutineNameEmpty)
+                _uiEvent.emit(InsertRoutineUiEvent.RoutineNameEmpty)
                 return@launch
             }
-            if (selectedDayOfWeek.isEmpty()) {
-                _uiEvent.emit(InsertRepeatRoutineUiEvent.InsertDayOfWeekEmpty)
-                return@launch
-            }
-            insertRepeatRoutineUseCase(
-                RepeatRoutine(
+            insertRoutineUseCase(
+                Routine(
+                    date = getTodayDate(),
                     text = routineName,
-                    dayOfWeekList = selectedDayOfWeek,
                     category = selectedCategory.value ?: ""
                 )
             )
-            _uiEvent.emit(InsertRepeatRoutineUiEvent.InsertSuccess)
+            _uiEvent.emit(InsertRoutineUiEvent.InsertSuccess)
         }
     }
 }
