+++ feature/insertRoutine/src/main/java/routiner/feature/insertRoutine/repeat/InsertRepeatRoutineRootScreen.kt
@@ -1,4 +1,4 @@
-package com.hegunhee.routiner.insertRoutine.repeat
+package routiner.feature.insertRoutine.repeat
 
 import android.content.Context
 import android.widget.Toast
@@ -38,15 +38,15 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.repeatOnLifecycle
-import com.hegunhee.routiner.insertRoutine.R
-import com.hegunhee.routiner.insertRoutine.common.CategoryDescriptionScreen
-import com.hegunhee.routiner.insertRoutine.common.CategoryLazyList
-import com.hegunhee.routiner.insertRoutine.common.CategoryTextEnterScreen
-import com.hegunhee.routiner.insertRoutine.common.DayOfWeekDescriptionScreen
-import com.hegunhee.routiner.insertRoutine.common.RoutineTextEnterScreen
+import routiner.feature.insertRoutine.common.CategoryDescriptionScreen
+import routiner.feature.insertRoutine.common.CategoryLazyList
+import routiner.feature.insertRoutine.common.CategoryTextEnterScreen
+import routiner.feature.insertRoutine.common.DayOfWeekDescriptionScreen
+import routiner.feature.insertRoutine.common.RoutineTextEnterScreen
 import routiner.core.model.DayOfWeek
 import routiner.core.ui.component.SelectableDayOfWeek
 import kotlinx.coroutines.flow.SharedFlow
+import routiner.feature.insertRoutine.R
 
 @Composable
 fun InsertRepeatRoutineRootScreen(
