+++ feature/insertRoutine/src/main/java/routiner/feature/insertRoutine/repeat/InsertRepeatRoutineViewModel.kt
@@ -1,56 +1,67 @@
-package com.hegunhee.routiner.insertRoutine.daily
+package routiner.feature.insertRoutine.repeat
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import routiner.core.domain.usecase.category.DeleteCategoryUseCase
 import routiner.core.domain.usecase.category.GetCategoriesFlowUseCase
 import routiner.core.domain.usecase.category.InsertCategoryUseCase
-import routiner.core.domain.usecase.routine.InsertRoutineUseCase
-import routiner.core.util.getTodayDate
+import routiner.core.domain.usecase.date.GetSortedDayOfWeekListUseCase
+import routiner.core.domain.usecase.repeatRoutine.InsertRepeatRoutineUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import routiner.core.model.Category
-import routiner.core.model.Routine
+import routiner.core.model.DayOfWeek
+import routiner.core.model.RepeatRoutine
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asSharedFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class InsertRoutineViewModel @Inject constructor(
+class InsertRepeatRoutineViewModel @Inject constructor(
     private val getCategoriesFlowUseCase: GetCategoriesFlowUseCase,
     private val deleteCategoryUseCase: DeleteCategoryUseCase,
     private val insertCategoryUseCase: InsertCategoryUseCase,
-    private val insertRoutineUseCase: InsertRoutineUseCase,
+    private val getSortedDayOfWeekListUseCase: GetSortedDayOfWeekListUseCase,
+    private val insertRepeatRoutineUseCase: InsertRepeatRoutineUseCase,
 ) : ViewModel() {
 
     private val categories = getCategoriesFlowUseCase()
 
     private val selectedCategory: MutableStateFlow<String?> = MutableStateFlow(null)
 
-    val uiState: StateFlow<InsertRoutineUiState> =
+    private val _dayOfWeek: MutableStateFlow<List<DayOfWeek>> =
+        MutableStateFlow(getSortedDayOfWeekListUseCase().map {
+            DayOfWeek(date = it, isSelected = false)
+        })
+    val dayOfWeek: StateFlow<List<DayOfWeek>> = _dayOfWeek.asStateFlow()
+
+    val uiState: StateFlow<InsertRepeatRoutineUiState> =
         categories.combine(selectedCategory) { categories, selectedCategory ->
             if (selectedCategory == null) {
-                InsertRoutineUiState.Categories(categories)
+                InsertRepeatRoutineUiState.Categories(categories, listOf())
             } else {
-                InsertRoutineUiState.Categories(categories.map { category ->
+                InsertRepeatRoutineUiState.Categories(categories.map { category ->
                     if (category.name == selectedCategory) category.copy(isSelected = true)
                     else category
-                })
+                }, listOf())
             }
+        }.combine(dayOfWeek) { categories, dayOfWeek ->
+            InsertRepeatRoutineUiState.Categories(categories.categories, dayOfWeek)
         }.stateIn(
             scope = viewModelScope,
-            initialValue = InsertRoutineUiState.Init,
+            initialValue = InsertRepeatRoutineUiState.Init,
             started = SharingStarted.WhileSubscribed(5_000)
         )
 
-    private val _uiEvent: MutableSharedFlow<InsertRoutineUiEvent> = MutableSharedFlow()
-    val uiEvent: SharedFlow<InsertRoutineUiEvent> = _uiEvent.asSharedFlow()
+    private val _uiEvent : MutableSharedFlow<InsertRepeatRoutineUiEvent> = MutableSharedFlow()
+    val uiEvent : SharedFlow<InsertRepeatRoutineUiEvent> = _uiEvent.asSharedFlow()
 
     fun onClickCategory(categoryName: String) {
         if (categoryName == selectedCategory.value) {
@@ -72,28 +83,38 @@ class InsertRoutineViewModel @Inject constructor(
     fun onClickCategoryInsert(insertCategoryText : String) {
         viewModelScope.launch {
             if(insertCategoryText.isBlank()) {
-                _uiEvent.emit(InsertRoutineUiEvent.InsertCategoryNameEmpty)
+                _uiEvent.emit(InsertRepeatRoutineUiEvent.InsertCategoryNameEmpty)
                 return@launch
             }
             insertCategoryUseCase(Category(insertCategoryText))
         }
-
     }
 
-    fun onClickRoutineInsert(routineName: String) {
+    fun onClickDayOfWeek(dayOfWeek : String) {
+        _dayOfWeek.value = _dayOfWeek.value.map {
+            if(it.date == dayOfWeek) it.copy(isSelected =  !it.isSelected)
+            else it
+        }
+    }
+
+    fun onClickRoutineInsert(routineName: String,selectedDayOfWeek : List<String>) {
         viewModelScope.launch {
             if (routineName.isBlank()) {
-                _uiEvent.emit(InsertRoutineUiEvent.RoutineNameEmpty)
+                _uiEvent.emit(InsertRepeatRoutineUiEvent.RoutineNameEmpty)
                 return@launch
             }
-            insertRoutineUseCase(
-                Routine(
-                    date = getTodayDate(),
+            if (selectedDayOfWeek.isEmpty()) {
+                _uiEvent.emit(InsertRepeatRoutineUiEvent.InsertDayOfWeekEmpty)
+                return@launch
+            }
+            insertRepeatRoutineUseCase(
+                RepeatRoutine(
                     text = routineName,
+                    dayOfWeekList = selectedDayOfWeek,
                     category = selectedCategory.value ?: ""
                 )
             )
-            _uiEvent.emit(InsertRoutineUiEvent.InsertSuccess)
+            _uiEvent.emit(InsertRepeatRoutineUiEvent.InsertSuccess)
         }
     }
 }
